feat(docker) Creating a common container for RAFT devs to use
This commit introduces a DockerFile that can be used to build an Ubuntu
22.04 or 24.04 image with all of RAFT's dependencies installed. The goal
of the image is to be able to run `bash raft/build.sh tests` with no
compilations errors.

We plan only to support ROCm 6.3 and beyond. Right now, no other ROCm
version is supported. We have included a build arg for it regardless, so
that we can avoid breaking anyone scripts when future ROCm versions are
supported by RAFT.