Fix(cmake) Update versions.json to correct format
The file cpp/build/install/lib/cmake/raft/raft-dependencies.cmake is generated
from CMake after calling `cmake --install .` or `make install` inside of a build
directory. It lists the dependencies that RAFT needs when you try to use it's
installation such as rmm, hipco, openmp, etc. Using find_package, a consuming
cmake project will look for all of rafts dependencies.

The names and details for each of these packages is defined in the versions.json
file that is not owned by raft (as far as I can tell). CMake takes the version
number from that json file and populates it in those find_package calls (maybe
elsewhere, but I don't know for sure). However, RAFT does own a versions.json file
which will overwrite from the main versions.json.

Cmake takes the values from raft/versions.json and writes the version number as
"branch-24.06" into the find package call, meaning the signature looks like this:
`find_package(hipco branch-24.06 QUIET)`. This line is then called when trying to
consume libraft via cmake, and causes installation to fail. This is because
"branch-24.06" is an invalid version format (contains letters).

This commit removes the entry all-together since version 0.0.1 and branch-24.06
is already the default in the main versions.json file. No reason to overwrite
an entry with the same thing.