Port headers required to enable DISTANCE_TEST (#42)
* Enable Distance Module tests

Get rid of all the inline GCN assembly in device_load_store_hip.cuh

The current implementaions were all missing the s_waitcnt instructions
which needs to be in place to correctly wait for all issued memory
operations to complete.

With this commit we get rid of all the inline assembly. The compiler
should automaticlly use flat addressing for the writes/reads to/from
LDS. Without knowing the actual perf hit by not using the ds_read/write
instructions it's better to go with something that's easier to get
right.

------------------------------------------------------------------------
Fix remaining tests due to thread block size and warp reductions being
performed in reference kernels
------------------------------------------------------------------------
Hipify the headers included by enabling the tests in gram.cu and
fused_cosine_nn.cu. This involved falling back to the non-CUTLAS kernel
when compiling with the HIP toolchain.