feat(src/random) Enable src/random
This commit enables src/random. Since the src gets compiled into
libraft.so, it is important that we test two things: we can link to
libraft.so by itself, and the raft_runtime APIs themselves.

Lastly, a test was introduced for the runtime random APIs. The test was copied
from the current random test and expanded for the datatypes instantiated
in the src API.

The build.sh file has been updated to re-introduce the --compile-lib options.

We are not testing the compile-static-lib option since only the .so is used
for building the python wheel, which is why enabling the src is prioritized.