refactor(tests) Move AMD test to their own directory (#52)
This commit moves the AMD tests to their own directory. This is
to make it easier to disambiguate which tests are AMD only and should
not be removed or altered in the case of a rebase. This also
lowers the diff when rebasing on upstream.

Also renamed the src test folder to more accurately represent which
portion of the src folder it is testing.