[Bugfix] Fix select_k mismatched index bug
This patch fixes the way how we index into the input distance array.
Previously we would use std::find to search through the whole of
"in_ids". This is incorrect as "in_inds" contains the input indices of
multiple batches whereas the input indices given to the comparator
"compare_ids" is batch specific.

Inputs to compare_ids:

```
0 <= i < spec.len
0 <= j < spec.len
in_inds = [                        # An array of size spec.len * spec.batch_size
  0, 1, 2, ..    (spec.len - 1),   # Batch:0
  0, 1, 2, ..    (spec.len - 1),   # Batch:1
  0, 1, 2, ..    (spec.len - 1),   # Batch:2
  ...
  0, 1, 2, ..    (spec.len - 1)    # Batch:spec.batch_size - 1
]
```
If we use std:find to search the whole of in_inds we will always match an
index that corresponds to Batch:0. When using this found index to query
into the input distance array "in_dist" will retrive an incorrect distance
value.

This custom comparison lambda was originally designed to allow mismatched
indices to be acceptable as long as the corresponding distances/values matched.

Due to the issue described above we were querying incorrect values/distances
and the above assumption could not be tested correctly.

This issue was first caught on an MI300X just by pure luck. The random seed
together with the input data generated with enough repeated values to hit
the above corner case.

The above bug also revealed other test cases where we weren't correctly
comparing indices/values. We currently add all the radix based select-k
algorithms to the forgive list; that is if inf/-inf values are involved
then we automatically pass the index comparison.