[COMMS_TEST] Add a test for the comms RAFT module (#19)
Add a new STD_COMMS_TEST test target. This test target
has a dependency on the raft::distributed interface
library which pulls in the UCX, MPI and RCCL
dependencies.

Add an MPI_COMMS_TEST when using the MPI
backend. Plumb the CMake infra to facilitate running
a CTest that invokes mpirun to launch multiple processes.