feat(linalg) Enable src
This commit enables the src/linalg. With this, raft_objs library is
complete, and we can begin work on enabling the python wheel.

Included is a test that verifies the API in src/linalg works as expected.
Most of the test code comes from the tests/linalg/coalesced_reduction.cu
file, but was expanded to cover all data types explicitly instantiated in
src/linalg/coalesced reduction.