fix(cmake) Fix spdlog install
The amd-integration is currently setup to use the rocmds-logger as
a third party dep. This is meant to hide the spdlog and fmt implementation
details and remove as a dep for install targets. However, hipMM still
uses an upstream version of RMM that must be packaged with spdlog/fmt
when installed.

A new branch has been created in hipMM that uses the new rocmds-logger
that avoids exporting fmt and spdlog.

This commit uses this new branch and is able to be consumed
without a spdlog or fmt install.