Fix excess_subsample (#2723)
A bug in the implementation could cause the resulting random samples to ignore the largest  ~5-10 percent of possible values.

This was identified to be the root cause for https://github.com/rapidsai/cuvs/issues/988 where the sub-sampling did not find representatives for a portion of samples.

This PR also adds a retry loop in case not enough unique values have been selected. Previously this could cause undefined results, although this is unlikely to happen and has not been reported yet.

CC @tfeher

Authors:
  - Malte Förster (https://github.com/mfoerste4)

Approvers:
  - Tamas Bela Feher (https://github.com/tfeher)

URL: https://github.com/rapidsai/raft/pull/2723