(sum/reduction) Fix intermittent summation tests
Due to the non-determnistic nature of how floating point reductions
are performed on the GPU we would see occasional failures of
`SumTests/SumTestD.Accuracy/4`. Every time this failure has happened
the absolute diff was around ~2.1. However taking into account
the scaled diff according to:

```c++
template <typename T>
struct CompareApprox {
  CompareApprox(T eps_) : eps(eps_) {}
  bool operator()(const T& a, const T& b) const
  {
    T diff  = std::abs(a - b);
    T m     = std::max(std::abs(a), std::abs(b));
    T ratio = diff > eps ? diff / m : diff;

    return (ratio <= eps);
  }

private:
  T eps;
};
```

The ratio we obtained was 1.5e-7. This was just above the tolerance
of 1e-7 for the failing test case.

This patch simply loosens the tolerance for failing test case from
1e-7 to 1e-5.