[AMD][fix] issue in lanczos_smallest
The second argument to `raft::linalg::gemv` is the following:

```
* @param [in] trans_a cublas transpose op for A. Type of trans_a is bool
```

The call site in `lanczos_smallest` was passing enum values `CUBLAS_OP_T`
and `CUBLAS_OP_N`. These enum values were being implicitly cast to bool.
On the CUDA platform `CUBLAS_OP_T` happened to be non-zero and `CUBLAS_OP_N`
is zero. This however isn't the case when using the corresponding values
defined in the hipBlas headers is:

```
typedef enum
{
    HIPBLAS_OP_N = 111, /**<  Operate with the matrix. */
    HIPBLAS_OP_T = 112, /**<  Operate with the transpose of the matrix. */
    HIPBLAS_OP_C = 113 /**< Operate with the conjugate transpose of the matrix. */
} hipblasOperation_t;
```

The fix here is to correctly pass true boolean values to indicate
whether to transpose or not.